# chore: improve coverage for options.key type