--- a/test/key.js
+++ b/test/key.js
@@ -5,6 +5,20 @@ const Fastify = require('fastify')
 const fastifySecureSession = require('..')
 const sodium = require('sodium-native')
 
+tap.test('throws when key is not string nor a Buffer', async t => {
+  t.plan(2)
+
+  const fastify = Fastify({ logger: false })
+  t.teardown(fastify.close.bind(fastify))
+
+  await fastify.register(fastifySecureSession, {
+    key: true
+  }).after(err => {
+    t.type(err, Error)
+    t.equal(err.message, 'key must be a string or a Buffer')
+  })
+})
+
 tap.test('support key length equals to "crypto_secretbox_KEYBYTES" length', t => {
   const fastify = Fastify({ logger: false })
   const key = Buffer.alloc(sodium.crypto_secretbox_KEYBYTES)