This PR improves test coverage.

<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [ ] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)

before:
```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 55.92ms
 PASS  test/cookie-name.js 6 OK 172.394ms
 PASS  test/key-base64.js 5 OK 175.017ms
 PASS  test/key-rotation.js 11 OK 185.281ms
 PASS  test/key.js 8 OK 197.355ms
 PASS  test/dynamic-cookie-options.js 13 OK 201.475ms
 PASS  test/delete.js 14 OK 183.093ms
 PASS  test/access.js 20 OK 211.846ms
 PASS  test/malformed-cookie.js 5 OK 151.069ms
 PASS  test/nonce-bad-length.test.js 5 OK 108.988ms
 PASS  test/path.js 4 OK 108.931ms
 PASS  test/schema.js 6 OK 126.143ms
 PASS  test/salt.js 5 OK 111.035ms
 PASS  test/secret.js 21 OK 2s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  132 passed, of 132
Time:   4s
ERROR: Coverage for lines (94.35%) does not meet global threshold (100%)
ERROR: Coverage for branches (92.18%) does not meet global threshold (100%)
ERROR: Coverage for statements (94.4%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|---------------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|---------------------------
All files |    94.4 |    92.18 |     100 |   94.35 |
 index.js |    94.4 |    92.18 |     100 |   94.35 | 70,81,118-119,141-142,262
----------|---------|----------|---------|---------|---------------------------
```

after:
```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 42.582ms
 PASS  test/delete.js 14 OK 173.648ms
 PASS  test/cookie-name.js 6 OK 170.049ms
 PASS  test/dynamic-cookie-options.js 13 OK 200.968ms
 PASS  test/key-rotation.js 11 OK 220.817ms
 PASS  test/key-base64.js 5 OK 178.558ms
 PASS  test/key.js 10 OK 199.973ms
 PASS  test/access.js 20 OK 220.745ms
 PASS  test/malformed-cookie.js 5 OK 129.342ms
 PASS  test/nonce-bad-length.test.js 5 OK 119.519ms
 PASS  test/path.js 4 OK 99.938ms
 PASS  test/schema.js 6 OK 133.054ms
 PASS  test/salt.js 5 OK 104.562ms
 PASS  test/secret.js 21 OK 2s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  134 passed, of 134
Time:   4s
ERROR: Coverage for lines (95.16%) does not meet global threshold (100%)
ERROR: Coverage for branches (93.75%) does not meet global threshold (100%)
ERROR: Coverage for statements (95.2%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|------------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|------------------------
All files |    95.2 |    93.75 |     100 |   95.16 |
 index.js |    95.2 |    93.75 |     100 |   95.16 | 81,118-119,141-142,262
----------|---------|----------|---------|---------|------------------------
```